LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/fetch-mixed-content-to-inscope.https-expected.txt

CONSOLE MESSAGE: FetchEvent.respondWith received an error: TypeError: Load failed
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?mode=no-cors&url=http%3A%2F%2Fweb-platform.test%3A8800%2Fservice-workers%2Fservice-worker%2Fresources%2Ffetch-access-control.py%3FPNGIMAGE.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: TypeError: Load failed
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?mode=no-cors&url=http%3A%2F%2Fwww1.web-platform.test%3A8800%2Fservice-workers%2Fservice-worker%2Fresources%2Ffetch-access-control.py%3FPNGIMAGE.


PASS Verify Mixed content of fetch() in a Service Worker
